+++ Source/Runtime/Metadata/Tables/FieldRVARow.cs
@@ -14,60 +14,60 @@ using System.IO;
 
 namespace Mosa.Runtime.Metadata.Tables 
 {
-    /// <summary>
-    /// 
-    /// </summary>
+	/// <summary>
+	/// 
+	/// </summary>
 	public struct FieldRVARow {
 		#region Data members
 
-        /// <summary>
-        /// 
-        /// </summary>
-        private uint _rva;
+		/// <summary>
+		/// 
+		/// </summary>
+		private ulong _rva;
 
-        /// <summary>
-        /// 
-        /// </summary>
-        private TokenTypes _fieldTableIdx;
+		/// <summary>
+		/// 
+		/// </summary>
+		private TokenTypes _fieldTableIdx;
 
-        #endregion // Data members
+		#endregion // Data members
 
-        #region Construction
+		#region Construction
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="FieldRVARow"/> struct.
-        /// </summary>
-        /// <param name="rva">The rva.</param>
-        /// <param name="fieldTableIdx">The field table idx.</param>
-        public FieldRVARow(uint rva, TokenTypes fieldTableIdx)
-        {
-            _rva = rva;
-            _fieldTableIdx = fieldTableIdx;
-        }
+		/// <summary>
+		/// Initializes a new instance of the <see cref="FieldRVARow"/> struct.
+		/// </summary>
+		/// <param name="rva">The rva.</param>
+		/// <param name="fieldTableIdx">The field table idx.</param>
+		public FieldRVARow(ulong rva, TokenTypes fieldTableIdx)
+		{
+			_rva = rva;
+			_fieldTableIdx = fieldTableIdx;
+		}
 
-        #endregion // Construction
+		#endregion // Construction
 
-        #region Properties
+		#region Properties
 
-        /// <summary>
-        /// Gets the rva.
-        /// </summary>
-        /// <value>The rva.</value>
-        public uint Rva
-        {
-            get { return _rva; }
-        }
+		/// <summary>
+		/// Gets the rva.
+		/// </summary>
+		/// <value>The rva.</value>
+		public ulong Rva
+		{
+			get { return _rva; }
+		}
 
 
-        /// <summary>
-        /// Gets the field table idx.
-        /// </summary>
-        /// <value>The field table idx.</value>
-        public TokenTypes FieldTableIdx
-        {
-            get { return _fieldTableIdx; }
-        }
+		/// <summary>
+		/// Gets the field table idx.
+		/// </summary>
+		/// <value>The field table idx.</value>
+		public TokenTypes FieldTableIdx
+		{
+			get { return _fieldTableIdx; }
+		}
 
-        #endregion // Properties
+		#endregion // Properties
 	}
 }
